backalmoustdone

+++ server/controllers/profile.controller.js
@@ -1,12 +1,10 @@
-const { Profile } = require("../models/models");
+const { User, Profile } = require("../models/models");
+// const { function } = require("joi");
 
 async function addProfile(req, res) {
   try {
     const { name, gender, birthdate, city } = req.body;
     const { id } = req.user;
-    if (!id) {
-      res.status(403).send({ message: "Not authorized" });
-    }
     const profile = await Profile.create({
       name,
       gender,
@@ -14,24 +12,41 @@ async function addProfile(req, res) {
       city,
       userId: id,
     });
+    const user = await User.findOne({ where: { id: id } });
+    user.update({ allProfiles: user.allProfiles + 1 });
     res.status(200).json({ profile });
   } catch (error) {
     res.status(404).send(error.message);
   }
 }
 
-async function getAllPrfoles(req, res) {
+async function getAllProfiles(req, res) {
   try {
-    const { id } = req.user;
-    if (!id) {
-      res.status(403).send({ message: "Not authorized" });
+    const {
+      params: { id },
+    } = req;
+    if (+id !== req.user.id) {
+      if (req.user.role !== "ADMIN") {
+        return res.status(401).send({ message: "You haven't rights" });
+      }
     }
     const getProfiles = await Profile.findAll({
+      where: { userId: id },
       attributes: ["id", "name", "gender", "birthdate", "city"],
     });
     res.status(200).json({ getProfiles });
   } catch (error) {
     res.status(404).send(error.message);
+  }
+}
+
+function check(id, user) {
+  if (+id === user.id) {
+    return id;
+  } else if (user.role === "ADMIN") {
+    return id;
+  } else {
+    return null;
   }
 }
 
@@ -45,8 +60,12 @@ async function updateProfile(req, res) {
     if (!profile) {
       return res.status(404).send({ message: "Profile not found" });
     }
-    profile.update({ name, gender, birthdate, city });
-    return res.status(200).json({ profile });
+    const checkId = check(profile.userId, req.user);
+    if (checkId) {
+      profile.update({ name, gender, birthdate, city });
+      return res.status(200).json({ profile });
+    }
+    res.status(401).send({ message: "Haven't rights" });
   } catch (error) {
     res.status(404).send(error.message);
   }
@@ -57,14 +76,26 @@ async function deleteProfile(req, res) {
     const {
       params: { id },
     } = req;
-    const deletedProfile = await Profile.destroy({ where: { id: id } });
-    if (!deletedProfile) {
-      return res.status(400).send("Profile not found");
+    const profile = await Profile.findOne({ where: { id: id } });
+    if (!profile) {
+      return res.status(404).send({ message: "Profile not found" });
     }
-    res.status(200).json({ message: "Profile was deleted" });
+    const checkId = check(profile.userId, req.user);
+    if (checkId) {
+      const deletedProfile = await Profile.destroy({ where: { id: id } });
+      const user = await User.findOne({ where: { id: profile.userId } });
+      user.update({ allProfiles: user.allProfiles - 1 });
+      return res.status(200).json({ message: "Profile was deleted" });
+    }
+    return res.status(401).send({ message: "You haven't rights" });
   } catch (error) {
     res.status(400).send(error.message);
   }
 }
 
-module.exports = { addProfile, getAllPrfoles, updateProfile, deleteProfile };
+module.exports = {
+  addProfile,
+  getAllProfiles,
+  updateProfile,
+  deleteProfile,
+};
